[DELETE ON CLUSTER PAGE]

+++ src/java/Servlet/ViewCluster.java
@@ -95,25 +95,27 @@ public class ViewCluster extends HttpServlet {
         User user = (User) request.getSession().getAttribute("loggedUser");
         Cluster cluster = (Cluster) request.getSession().getAttribute("clusterToProcess");
         CircaDatabase db = CircaDatabase.getInstance();
-            
-        if(type.equals("add-cluster-member")){
-            String[] newmember = request.getParameterValues("new-member");
-            if (newmember != null) {
-                for (String s : newmember) {
-                    db.addUserToCluster(user.getUserID(), Integer.parseInt(s), cluster.getClusterID());
+        
+        if(type != null){
+            if(type.equals("add-cluster-member")){
+                String[] newmember = request.getParameterValues("new-member");
+                if (newmember != null) {
+                    for (String s : newmember) {
+                        db.addUserToCluster(user.getUserID(), Integer.parseInt(s), cluster.getClusterID());
+                    }
                 }
             }
-        }
-        else if(type.equals("delete-cluster-member")){
-            int clusterMemberID = Integer.parseInt(request.getParameter("cluster-member-id"));
-            
-            db.deleteUsertoCluster(clusterMemberID, cluster.getClusterID());
-        }
-        else if(type.equals("edit-cluster-name")){
-            String name = request.getParameter("cluster-name");
-            int clusterID = Integer.parseInt(request.getParameter("cluster-id"));
-            cluster.setName(name);
-            db.editClusterName(clusterID, name);
+            else if(type.equals("delete-cluster-member")){
+                int clusterMemberID = Integer.parseInt(request.getParameter("cluster-member-id"));
+
+                db.deleteUsertoCluster(clusterMemberID, cluster.getClusterID());
+            }
+            else if(type.equals("edit-cluster-name")){
+                String name = request.getParameter("cluster-name");
+                int clusterID = Integer.parseInt(request.getParameter("cluster-id"));
+                cluster.setName(name);
+                db.editClusterName(clusterID, name);
+            }
         }
         RequestDispatcher reqDispatcher = request.getRequestDispatcher("ClusterPage.jsp");
             reqDispatcher.forward(request, response);
